software-data-collection/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
airtable==0.4.8
amqp==5.1.1
asgiref==3.7.2
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
billiard==4.1.0
boto3==1.26.158
botocore==1.29.158
cachetools==5.3.1
celery==5.3.1
certifi==2023.5.7
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cmake==3.27.0
distlib==0.3.6
Django==4.2.2
django-cors-headers==4.1.0
djangorestframework==3.14.0
exceptiongroup==1.1.1
filelock==3.12.2
flake8==6.0.0
flower==2.0.0
frozenlist==1.3.3
fsspec==2023.6.0
google-auth==2.20.0
google-auth-oauthlib==1.0.0
gspread==5.9.0
h11==0.14.0
httplib2==0.22.0
huggingface-hub==0.16.4
humanize==4.6.0
identify==2.5.24
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.1
kombu==5.3.1
langdetect==1.0.9
lit==16.0.6
lxml==4.9.2
MarkupSafe==2.1.3
mccabe==0.7.0
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
nltk==3.8.1
nodeenv==1.8.0
numpy==1.25.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauth2client==4.1.3
oauthlib==3.2.2
openai==0.27.8
outcome==1.2.0
packaging==23.1
Pillow==10.0.0
platformdirs==3.8.0
pre-commit==3.3.3
prometheus-client==0.17.0
prompt-toolkit==3.0.38
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pyflakes==3.0.1
pyparsing==3.1.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
random-user-agent==1.0.1
redis==4.5.5
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.1
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
selenium==4.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
sympy==1.12
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
transformers==4.31.0
trio==0.22.0
trio-websocket==0.10.3
triton==2.0.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
uWSGI==2.0.21
vine==5.0.0
virtualenv==20.23.1
wcwidth==0.2.6
wsproto==1.2.0
yarl==1.9.2
